refactor(archtest): SAGA-PROJECTION-DEPS-INMEM-FUNNEL-01 并入 saga 统一文件 [#2175]
把 sagaprojection_deps_inmem_funnel{,_test}.go 折叠进 saga 主题统一双文件
（helper → saga_invariants.go，test → saga_invariants_test.go），删除两个独立
文件，消解 TestSagaInvariantsConsolidated（SAGA-INVARIANTS-FILE-CONSOLIDATED-01）
对 stray saga invariant 文件的报红。

- saga_invariants_test.go: invariants header / merged-from note / knownSagaInvariantIDs
  三处同步（新 ID 纳入 B1 oracle，自动抬升非-vacuity 楼层）。
- saga_invariants.go: 花名册加 CheckSagaProjectionDepsInmemFunnel01；顺手去掉已陈旧
  的 "ten" 计数措辞（改为 count-agnostic，de-brittle）。
- 规则 ID / 评级 / anti-vacuity 与 RED-fixture / dot-import 盲区测试整块保留。
